clients/cli-chatbot/pyproject.toml
[project]
name = "cli-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "anthropic>=0.56.0",
    "google-genai>=1.24.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.16",
    "langchain-core>=0.3.67",
    "langchain-google-genai>=2.0.10",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.27",
    "mcp[cli]>=1.9.4",
    "ollama>=0.5.1",
    "python-dotenv>=1.1.0",
]
